Revert "feat: rebuild noblocks homepage (#149)"
This reverts commit 664f96b00a4043740e9bdf4b3e39eb8867799445.
