+++ app/context/StepContext.tsx
@@ -1,4 +1,3 @@
-"use client"
 import React, { createContext, useContext, useState } from "react";
 import { Step, STEPS } from "../types";
