+++ app/layout.tsx
@@ -209,7 +209,7 @@ export default function RootLayout({
         <Providers>
           <div className="min-h-full min-w-full bg-white transition-colors dark:bg-neutral-900">
             <Navbar />
-            <div className="relative mx-auto flex min-h-dvh flex-col items-center pt-24 transition-all">
+            <div className="relative mx-auto flex min-h-dvh flex-col items-center px-5 pt-24 transition-all">
               <MainContent>{children}</MainContent>
               <Footer />
             </div>
