+++ next.config.mjs
@@ -34,9 +34,6 @@ const nextConfig = {
   compiler: {
     removeConsole: process.env.NODE_ENV === "production",
   },
-  images: {
-    domains: ["flagcdn.com"],
-  },
 };
 
 export default nextConfig;
